+++ projects/women-in-tech/DistributionGraphic.jsx
@@ -1,14 +1,8 @@
 import * as React from "react";
 import PropTypes from "prop-types";
-import {
-  DistributionContainer,
-  DistributionFooter,
-  DistributionMeasures,
-  DistributionTitle,
-  CareerEntry
-} from "./graphic-fragments";
-import { buildPath, GraphicConstants, getLastItem, getVerticalPointFromDomain } from "../utils";
-import { generateSlots, getFilledSlots } from "../slotUtils";
+import Graphic from "./Graphic/index";
+import { buildPath, GraphicConstants, getLastItem, getVerticalPointFromDomain } from "./utils";
+import { generateSlots, getFilledSlots } from "./slotUtils";
 
 const { BOTTOM_LIMIT, TOP_LIMIT } = GraphicConstants;
 
@@ -25,7 +19,7 @@ const DistributionGraphic = ({ data, ...extraProps }) => {
     config: { yDomain, title, legend }
   } = graphicData;
 
-  const onCareerClickHandler = (careerName) => {
+  const onCareerClickHandler = React.useCallback((careerName) => {
     setGraphicData((prev) => ({
       ...prev,
       entries: prev.entries.map((career) =>
@@ -34,7 +28,7 @@ const DistributionGraphic = ({ data, ...extraProps }) => {
           : { ...career, highlight: false, path: { ...career.path, className: "st1" } }
       )
     }));
-  };
+  }, []);
 
   const filledSlots = React.useMemo(() => {
     const slots = generateSlots({
@@ -63,8 +57,8 @@ const DistributionGraphic = ({ data, ...extraProps }) => {
   }, [entries]);
 
   return (
-    <DistributionContainer {...extraProps}>
-      <DistributionTitle title={title} />
+    <Graphic {...extraProps}>
+      <Graphic.Title title={title} />
 
       {/* Entries Evolution + Career */}
       {entries.map((entry) => {
@@ -73,7 +67,7 @@ const DistributionGraphic = ({ data, ...extraProps }) => {
         let lastPointY = filledSlots[entry.text.value];
 
         return (
-          <CareerEntry
+          <Graphic.Entry
             key={entry.text.value}
             highlight={entry.highlight}
             pathData={{ path, ...pathData }}
@@ -83,9 +77,9 @@ const DistributionGraphic = ({ data, ...extraProps }) => {
         );
       })}
 
-      <DistributionMeasures measures={measures} legend={legend} />
-      <DistributionFooter />
-    </DistributionContainer>
+      <Graphic.Measure measures={measures} legend={legend} />
+      <Graphic.Footer />
+    </Graphic>
   );
 };
